dcncel button

+++ server.js
@@ -606,16 +606,38 @@ slapp.action('cancel_request', 'cancel', (msg, value) => {
   var email = arr[0]
   var vacationId = arr[1]
   toffyHelper.getNewSessionwithCookie(email, function (remember_me_cookie, session_Id) {
+    //get vacation state
+    var uri = 'http://' + IP + '/api/v1/vacation/' + vacationId
     request({
-      url: 'http://' + IP + '/api/v1/vacation/' + vacationId,
-      method: 'DELETE',
+      url: uri, //URL to hitDs
+      method: 'GET',
       headers: {
         'Content-Type': 'application/json',
-        'Cookie': remember_me_cookie + ";" + session_Id
-      },
+        'Cookie': remember_me_cookie + ";" + session_id
+
+      }
+      //Set the body as a stringcc
     }, function (error, response, body) {
-      msg.respond(msg.body.response_url, "Your request has been canceled")
+      if ((JSON.parse(body)).vacationState == "PendingManagersApproval") {
+        //delete vacation request
+        request({
+          url: 'http://' + IP + '/api/v1/vacation/' + vacationId,
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+            'Cookie': remember_me_cookie + ";" + session_Id
+          },
+        }, function (error, response, body) {
+          msg.respond(msg.body.response_url, "Your request has been canceled")
+        })
+      } else {
+        //the managers take an action
+        msg.respond(msg.body.response_url, "Sorry ,you can't cancel your time off request ,since your managers take an action.Please contact them")
+
+
+      }
     })
+
   })
 })
